requirements.txt
streamlit>=1.35
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
seaborn>=0.13
scikit-learn>=1.4
xgboost>=2.0
python-dotenv>=1.0
openai>=1.30
